+++ src/main/java/org/akcjamis/webapp/repository/ChristmasPackageChangeRepository.java
@@ -6,14 +6,12 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.*;
 
-import java.util.List;
-
 /**
  * Spring Data JPA repository for the ChristmasPackageChange entity.
  */
-public interface ChristmasPackageChangeRepository extends JpaRepository<ChristmasPackageChange,Long> {
+public interface ChristmasPackageChangeRepository extends JpaRepository<ChristmasPackageChange, Integer> {
 
-    Page<ChristmasPackageChange> findByChristmasPackage_id(Long id, Pageable var1);
+    Page<ChristmasPackageChange> findByChristmasPackage_id(Integer id, Pageable var1);
 
-    ChristmasPackageChange findByIdAndChristmasPackage_id(Long packageId, Long id);
+    ChristmasPackageChange findByIdAndChristmasPackage_id(Integer packageId, Integer id);
 }
