+++ src/main/java/org/akcjamis/webapp/repository/SubpackageNoteRepository.java
@@ -4,11 +4,9 @@ import org.akcjamis.webapp.domain.SubpackageNote;
 
 import org.springframework.data.jpa.repository.*;
 
-import java.util.List;
-
 /**
  * Spring Data JPA repository for the SubpackageNote entity.
  */
-public interface SubpackageNoteRepository extends JpaRepository<SubpackageNote,Long> {
+public interface SubpackageNoteRepository extends JpaRepository<SubpackageNote, Integer> {
 
 }
