+++ src/main/java/org/akcjamis/webapp/repository/SubpackageRepository.java
@@ -4,11 +4,9 @@ import org.akcjamis.webapp.domain.Subpackage;
 
 import org.springframework.data.jpa.repository.*;
 
-import java.util.List;
-
 /**
  * Spring Data JPA repository for the Subpackage entity.
  */
-public interface SubpackageRepository extends JpaRepository<Subpackage,Long> {
+public interface SubpackageRepository extends JpaRepository<Subpackage, Integer> {
 
 }
