+++ src/main/java/org/akcjamis/webapp/repository/search/FamilyNoteSearchRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
 /**
  * Spring Data ElasticSearch repository for the FamilyNote entity.
  */
-public interface FamilyNoteSearchRepository extends ElasticsearchRepository<FamilyNote, Long> {
+public interface FamilyNoteSearchRepository extends ElasticsearchRepository<FamilyNote,Integer> {
 }
