+++ src/main/java/org/akcjamis/webapp/repository/search/SubpackageSearchRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
 /**
  * Spring Data ElasticSearch repository for the Subpackage entity.
  */
-public interface SubpackageSearchRepository extends ElasticsearchRepository<Subpackage, Long> {
+public interface SubpackageSearchRepository extends ElasticsearchRepository<Subpackage, Integer> {
 }
