+++ src/main/java/org/akcjamis/webapp/repository/search/TeamSearchRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
 /**
  * Spring Data ElasticSearch repository for the Team entity.
  */
-public interface TeamSearchRepository extends ElasticsearchRepository<Team, Long> {
+public interface TeamSearchRepository extends ElasticsearchRepository<Team, Integer> {
 }
